Refactor BenchCel into a single device class
